Release 279 to merge w develop

+++ integration-test/src/test/java/org/sagebionetworks/ITOpenIDConnectTest.java
@@ -3,9 +3,13 @@ package org.sagebionetworks;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
+import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.Collections;
+import java.util.List;
+import java.util.Map;
 
+import org.json.JSONArray;
 import org.json.JSONObject;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.AfterEach;
@@ -16,6 +20,7 @@ import org.sagebionetworks.client.SynapseAdminClientImpl;
 import org.sagebionetworks.client.SynapseClient;
 import org.sagebionetworks.client.SynapseClientImpl;
 import org.sagebionetworks.client.exceptions.SynapseException;
+import org.sagebionetworks.repo.model.UserProfile;
 import org.sagebionetworks.repo.model.auth.JSONWebTokenHelper;
 import org.sagebionetworks.repo.model.oauth.JsonWebKeySet;
 import org.sagebionetworks.repo.model.oauth.OAuthAuthorizationResponse;
@@ -24,6 +29,7 @@ import org.sagebionetworks.repo.model.oauth.OAuthClientIdAndSecret;
 import org.sagebionetworks.repo.model.oauth.OAuthClientList;
 import org.sagebionetworks.repo.model.oauth.OAuthGrantType;
 import org.sagebionetworks.repo.model.oauth.OAuthResponseType;
+import org.sagebionetworks.repo.model.oauth.OAuthScope;
 import org.sagebionetworks.repo.model.oauth.OIDCAuthorizationRequest;
 import org.sagebionetworks.repo.model.oauth.OIDCAuthorizationRequestDescription;
 import org.sagebionetworks.repo.model.oauth.OIDCSigningAlgorithm;
@@ -34,7 +40,7 @@ import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwsHeader;
 import io.jsonwebtoken.Jwt;
 
-class ITOpenIDConnectTest {
+public class ITOpenIDConnectTest {
 
 	private static SynapseAdminClient adminSynapse;
 	private static SynapseClient synapseOne;
@@ -63,10 +69,10 @@ class ITOpenIDConnectTest {
 	@AfterAll
 	public static void afterClass() throws Exception {
 		try {
-			adminSynapse.deleteUser(user1ToDelete);
+			if (user1ToDelete!=null) adminSynapse.deleteUser(user1ToDelete);
 		} catch (SynapseException e) { }
 		try {
-			adminSynapse.deleteUser(user2ToDelete);
+			if (user2ToDelete!=null) adminSynapse.deleteUser(user2ToDelete);
 		} catch (SynapseException e) { }
 	}
 
@@ -83,7 +89,7 @@ class ITOpenIDConnectTest {
 
 	
 	@Test
-	void testRoundTrip() throws Exception {
+	public void testRoundTrip() throws Exception {
 		OIDConnectConfiguration connectConfig = synapseAnonymous.getOIDConnectConfiguration();
 		assertNotNull(connectConfig.getIssuer());
 		
@@ -96,8 +102,7 @@ class ITOpenIDConnectTest {
 		client = synapseOne.createOAuthClient(client);
 		clientToDelete = client.getClient_id();
 		
-		OAuthClient retrieved = synapseOne.getOAuthClient(client.getClient_id());
-		assertEquals(client, retrieved);
+		assertEquals(client, synapseOne.getOAuthClient(client.getClient_id()));
 		
 		OAuthClientList clientList = synapseOne.listOAuthClients(null);
 		assertEquals(client, clientList.getResults().get(0));
@@ -111,7 +116,10 @@ class ITOpenIDConnectTest {
 		authorizationRequest.setRedirectUri(client.getRedirect_uris().get(0));
 		authorizationRequest.setResponseType(OAuthResponseType.code);
 		authorizationRequest.setScope("openid");
-		authorizationRequest.setClaims("{\"id_token\":{\"userid\":\"null\"},\"userinfo\":{\"userid\":\"null\"}}");
+		authorizationRequest.setClaims(
+				"{\"id_token\":{\"userid\":\"null\",\"email\":null,\"is_certified\":null,\"team\":{\"values\":[\"2\"]}},"+
+				 "\"userinfo\":{\"userid\":\"null\",\"email\":null,\"is_certified\":null,\"team\":{\"values\":[\"2\"]}}}"
+		);
 		
 		// Note, we get the authorization description anonymously
 		OIDCAuthorizationRequestDescription description = 
@@ -132,19 +140,42 @@ class ITOpenIDConnectTest {
 		}
 		
 		Jwt<JwsHeader, Claims> parsedIdToken = JSONWebTokenHelper.parseJWT(tokenResponse.getId_token(), jsonWebKeySet);
-		String myId = synapseOne.getMyProfile().getOwnerId();
-		assertEquals(myId, parsedIdToken.getBody().get("userid", String.class));
+		UserProfile myProfile = synapseOne.getMyProfile();
+		String myId = myProfile.getOwnerId();
+		Claims idClaims = parsedIdToken.getBody();
+		assertEquals(myId, idClaims.get("userid", String.class));
+		assertTrue(idClaims.get("is_certified", Boolean.class));
+		String email = myProfile.getEmails().get(0);
+		assertEquals(email, idClaims.get("email", String.class));
+		assertEquals(Collections.EMPTY_LIST, idClaims.get("team", List.class));
+		
+		// the access token encodes claims we can refresh
+		Jwt<JwsHeader, Claims> parsedAccessToken = JSONWebTokenHelper.parseJWT(tokenResponse.getAccess_token(), jsonWebKeySet);
+		Claims accessClaims = parsedAccessToken.getBody();
+		Map access = (Map)accessClaims.get("access", Map.class);
+		List<String> userInfoScope = (List<String>)access.get("scope");
+		assertEquals(1, userInfoScope.size());
+		assertEquals(OAuthScope.openid.name(), userInfoScope.get(0));
+		Map userInfoClaims = (Map)access.get("oidc_claims");
+		assertTrue(userInfoClaims.containsKey("userid"));
+		assertTrue(userInfoClaims.containsKey("email"));
+		assertTrue(userInfoClaims.containsKey("is_certified"));
+		assertTrue(userInfoClaims.containsKey("team"));
 
 		// Note, we use a bearer token to authorize the client 
 		try {
 			synapseAnonymous.setBearerAuthorizationToken(tokenResponse.getAccess_token());
 			JSONObject userInfo = synapseAnonymous.getUserInfoAsJSON();
-			// spot check verify userInfo
-			assertEquals(myId, (String)userInfo.get("userid"), userInfo.toString());
+			// check userInfo
+			assertEquals(myId, (String)userInfo.get("userid"));
+			assertEquals(email, (String)userInfo.get("email"));
+			assertTrue((Boolean)userInfo.get("is_certified"));
+			assertEquals(0, ((JSONArray)userInfo.get("team")).length());
 		} finally {
 			synapseAnonymous.removeAuthorizationHeader();
 		}
 		
+		OAuthClient retrieved = synapseOne.getOAuthClient(client.getClient_id());
 		retrieved.setUserinfo_signed_response_alg(OIDCSigningAlgorithm.RS256);
 		OAuthClient updated = synapseOne.updateOAuthClient(retrieved);
 		assertEquals(retrieved.getClient_name(), updated.getClient_name());
@@ -153,7 +184,11 @@ class ITOpenIDConnectTest {
 		try {
 			synapseAnonymous.setBearerAuthorizationToken(tokenResponse.getAccess_token());
 			Jwt<JwsHeader,Claims> userInfo = synapseAnonymous.getUserInfoAsJSONWebToken();
-			assertEquals(myId, userInfo.getBody().get("userid", String.class));
+			Claims body = userInfo.getBody();
+			assertEquals(myId, body.get("userid", String.class));
+			assertEquals(email, body.get("email", String.class));
+			assertTrue(body.get("is_certified", Boolean.class));
+			assertEquals(Collections.EMPTY_LIST, body.get("team", List.class));
 		} finally {
 			synapseAnonymous.removeAuthorizationHeader();
 		}
